refactor(frontend): decompose AlertDialog, restore hydratedPickRef fix (spec 14, partial)
- extract dialog/types.ts (CarriedWidget + wire shapes), dialog/state.ts
  (ymd/parseCron/buildCron/defaultAggFromSpec), dialog/FilterDetails.tsx
  (Number/DateRange/Values pickers)
- AlertDialog.tsx 2486 -> 2241 lines; as any 164 -> 132; useState 102 -> 92
- re-introduces the stashed hydratedPickRef fix: saved render.mode no longer
  clobbered by async widget-cfg load; Attach-PDF row shows when report OR attachPdf
- default export + prop signature stable (widget widened to CarriedWidget superset)

Safe partial: useReducer + section-component split (steps 4-5) and full as-any
sweep deferred (need AlertConfig narrowing in lib/api.ts, out of partition).

Spec: v2 Roadmap/P1-frontend/14
